zadanie3.6

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -4,15 +4,15 @@ import static java.lang.Math.sqrt;
 
 public class Calculator {
 
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public int sub(int a, int b) {
+    public double sub(double a, double b) {
         return a - b;
     }
 
-    public int sqr(double a) {
-        return (int) Math.pow(a, 2);
+    public double sqr(double a) {
+        return Math.pow(a,2);
     }
 }
